Alpha v2.0

+++ s2e_env/commands/init.py
@@ -1,6 +1,4 @@
 """
-MIT License
-
 Copyright (c) 2017 Dependable Systems Laboratory, EPFL
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -30,12 +28,12 @@ import sys
 
 import requests
 import sh
-from sh import git, ErrorReturnCode
+from sh import ErrorReturnCode
 from sh.contrib import sudo
 
 from s2e_env import CONSTANTS
 from s2e_env.command import BaseCommand, CommandError
-
+from s2e_env.utils import repos
 
 class Command(BaseCommand):
     """
@@ -50,14 +48,32 @@ class Command(BaseCommand):
         parser.add_argument('dir', help='The environment directory')
         parser.add_argument('-s', '--skip-dependencies', action='store_true',
                             help='Skip the dependency install via apt')
+        parser.add_argument('-b', '--use-existing-install', required=False, default=None,
+                            help='Do not fetch sources but instead use existing S2E installation '
+                                 'whose prefix is specified by this parameter (e.g., /opt/s2e)')
         parser.add_argument('-f', '--force', action='store_true',
                             help='Use this flag to force environment creation '
                                  'even if an environment already exists at '
                                  'this location')
 
-    def handle(self, **options):
+    def install_binary_dist(self, env_path, prefix):
+        # We must use an absolute path because of symlinks
+        prefix = os.path.abspath(prefix)
+
+        self.info('Using S2E installation in %s' % prefix)
+        install = os.path.join(env_path, 'install')
+        shutil.rmtree(install, True)
+        os.symlink(prefix, install)
+
+        # We still need to clone guest-images repo, because it contains
+        # info about location of the images
+        guest_images_repo = CONSTANTS['repos']['images']['build']
+        repos.git_clone_to_source(env_path, guest_images_repo)
+
+    def handle(self, *args, **options):
         env_path = os.path.realpath(options['dir'])
         force = options['force']
+        prefix = options['use_existing_install']
 
         # First check if we are already in the environment directory
         if os.path.realpath(env_path) == os.path.realpath(os.getcwd()):
@@ -65,7 +81,7 @@ class Command(BaseCommand):
                                'current working directory. Please `cd ..`')
 
         # Then check if something already exists at the environment directory
-        if os.path.isdir(env_path):
+        if os.path.isdir(env_path) and not os.listdir(env_path) == []:
             if force:
                 self.info('%s already exists - removing' % env_path)
                 shutil.rmtree(env_path)
@@ -78,25 +94,33 @@ class Command(BaseCommand):
                                    '``s2e build`` or ``s2e update`` instead' %
                                    env_path)
 
-        # Install S2E's dependencies via apt-get
-        if not options['skip_dependencies']:
-            self._install_dependencies()
 
-        # At this point it is safe to assume that the environment directory
-        # does not exist. So let's create it
+        # Create environment if it doesn't exist
         self.info('Creating environment in %s' % env_path)
-        os.mkdir(env_path)
+        if not os.path.isdir(env_path):
+            os.mkdir(env_path)
 
         # Create the directories within the environment
         for dir_ in CONSTANTS['dirs']:
             os.mkdir(os.path.join(env_path, dir_))
 
-        # Get the source repositories
-        self._get_s2e_sources(env_path)
-        self._get_img_sources(env_path)
+        # Create marker file to recognize s2e environment folder
+        with open(os.path.join(env_path, '.s2eenv'), 'w'):
+            pass
 
-        return ('Environment created in %s. Now run ``s2e build`` to build' %
-                env_path)
+        if prefix is not None:
+            self.install_binary_dist(env_path, prefix)
+            return 'Environment created in %s.' % env_path
+        else:
+            # Install S2E's dependencies via apt-get
+            if not options['skip_dependencies']:
+                self._install_dependencies()
+
+            # Get the source repositories
+            self._get_s2e_sources(env_path)
+            self._get_img_sources(env_path)
+
+            return 'Environment created in %s. Now run ``s2e build`` to build' % env_path
 
     def _install_dependencies(self):
         """
@@ -186,22 +210,10 @@ class Command(BaseCommand):
         """
         Download the S2E image repositories.
         """
-        git_url = CONSTANTS['repos']['url']
         git_repos = CONSTANTS['repos']['images'].values()
 
         for git_repo in git_repos:
-            git_repo_dir = os.path.join(env_path, 'source', git_repo)
-            git_repo_url = '%s/%s' % (git_url, git_repo)
-
-            try:
-                self.info('Fetching %s from %s' % (git_repo, git_repo_url))
-                git.clone(git_repo_url, git_repo_dir, _out=sys.stdout,
-                          _err=sys.stderr, _fg=True)
-            except ErrorReturnCode as e:
-                raise CommandError(e)
-
-            # Success!
-            self.success('Fetched %s' % git_repo)
+            repos.git_clone_to_source(env_path, git_repo)
 
     def _get_repo(self, env_path):
         """
@@ -209,7 +221,7 @@ class Command(BaseCommand):
 
         If the repo binary does not exist, download it.
         """
-        repo_path = os.path.join(env_path, 'bin', 'repo')
+        repo_path = os.path.join(env_path, 'install', 'bin', 'repo')
         if not os.path.isfile(repo_path):
             self._download_repo(repo_path)
